refactor: move oc10 related functions
